Added imagick as requirement to the installer

+++ src/setup/tests/check-php-settings.php
@@ -18,5 +18,10 @@ if (!class_exists("\\finfo")) {
 	echo json_encode(array("error" => true, "errormessage" => "You are missing the fileinfo-library for PHP. Please install and activate it."));
 	exit;
 }
+
+if (!class_exists("\\Imagick")) {
+	echo json_encode(array("error" => true, "errormessage" => "You are missing the Imagick library for PHP. Please install and activate it."));
+	exit;
+}
 echo json_encode(array("error" => false));
 exit;
